requirements.txt
# --- Core RAG & LangChain Framework ---
langchain
langchain-core
langchain-community
langchain-text-splitter
langchain-chroma

# --- LLM & Embedding Model Integrations ---
langchain-google-genai
langchain-openai
azure-identity

# --- Web Application ---
streamlit

# --- Data Scraping & Processing ---
requests
beautifulsoup4
pandas
unstructured[local-inference] # Provides loaders for various file types like .txt, .pdf, etc.
wikipedia

# --- Configuration ---
python-dotenv
